requirements.txt
discord.py
python-dotenv
motor
dnspython
certifi